% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marecats_functions.R
\name{kitty_mood_enrichment}
\alias{kitty_mood_enrichment}
\title{Kitty Mood Enrichment}
\usage{
kitty_mood_enrichment(weather_condition, current_mood)
}
\arguments{
\item{weather_condition}{A character string representing the current weather.
Valid values include "rainy", "hot", "snowy", "cold", "sunny", "warm", "mild", "nice".}

\item{current_mood}{An integer of the current mood of the cat, ranging from 1 (low mood) to 10 (high mood).}
}
\value{
A message summarizing the cat's mood and suggesting enrichment activities to potentially improve the mood.
}
\description{
This function calculates the cat's mood based on weather conditions
and suggests enrichment activities to improve the mood.
}
\examples{
kitty_mood_enrichment("snowy", 1)
}
