% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marecats_functions.R
\name{kitty_treats}
\alias{kitty_treats}
\title{Kitty Treats Calculation}
\usage{
kitty_treats(last_fed)
}
\arguments{
\item{last_fed}{A character string representing the last feeding time in "HH:MM" format.
This should be the time when the cat was last fed on the current day.}
}
\value{
A character string suggesting the number of treats to give.
}
\description{
This function calculates how many treats to give a cat based on the time since the last feeding.
The calculation is based on one treat per hour.
}
\examples{
kitty_treats("11:34")
}
